Updated HSbotRunner.bat, reorganization
The batch script now checks if there has been a timestamp of the last update of lib requirements. If there is no timestamp or it's older than 30 days it installs the requirements for MFB to run otherwise it does not, this will shave seconds off of each launch. %~dp0 variable was also removed to prevent further errors as all file references are from current working directory.
Developer files were moved to \dev
Shell script with new changes coming next.